refactor(database): reorganize database services folder into domain-based subfolders
