+++ packages/database/src/services/quotes/quotes.service.test.ts
@@ -7,7 +7,7 @@ import {
   mockReturning,
   mockFindFirst,
   mockFindMany,
-} from "../tests/utils/db-mock";
+} from "../../tests/utils/db-mock";
 import { QuotesService, type ComplexQuote } from "./quotes.service";
 import {
   type TQuote,
@@ -15,8 +15,8 @@ import {
   type TQuoteMessage,
   type TUser,
   type TProduct,
-} from "../schemas";
-import type { IDatabase } from "../client";
+} from "../../schemas";
+import type { IDatabase } from "../../client";
 
 const quotesService = new QuotesService(mockDb as unknown as IDatabase);
