+++ packages/database/src/services/warehouse-stock/warehouse-stock.service.ts
@@ -1,9 +1,9 @@
 import { eq, sql } from "drizzle-orm";
-import type { IDatabase } from "../client";
-import { warehouseStocks } from "../schemas/warehouse-stock.schema";
-import { products } from "../schemas/product.schema";
-import type { IWarehouseStockService } from "./interfaces";
-import type { TUpdateWarehouseStockInput } from "../validators";
+import type { IDatabase } from "../../client";
+import { warehouseStocks } from "../../schemas/warehouse-stock.schema";
+import { products } from "../../schemas/product.schema";
+import type { IWarehouseStockService } from "../interfaces";
+import type { TUpdateWarehouseStockInput } from "../../validators";
 
 export class WarehouseStockService implements IWarehouseStockService {
   constructor(protected readonly db: IDatabase) {}
