+++ process/general-plans/completed/database-services-refactor_PLAN_12-06-26.md
@@ -4,7 +4,7 @@
 **Complexity**: SIMPLE
 **Implementation Approach**: Domain-based Subfolders Reorganization
 **Execution Model**: Folder Reorganization followed by Interface/Registry Path Mapping
-**Status**: ⏳ PLANNED (Backlog)
+**Status**: ✅ VERIFIED
 
 ---
 
@@ -127,4 +127,4 @@ All unit and integration tests must pass successfully.
 2. Start with **Phase 1: Folder Reorganization** by moving service files and tests into subfolders.
 3. Validate by executing the check-types and test runner verification gates.
 
-**Next Step**: Review this plan, approve for execution, then enter EXECUTE mode.
+**Next Step**: None. Implementation and verification completed.
